requirements.txt
aiofiles==0.4.0
Pillow==7.0.0
pyzbar==0.1.8
